+++ docs/classes/lines.html
@@ -145,7 +145,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/services/LineService.ts#L215">services/LineService.ts:215</a></li>
+									<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/services/LineService.ts#L215">services/line.service.ts:215</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -184,7 +184,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/services/LineService.ts#L87">services/LineService.ts:87</a></li>
+									<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/services/LineService.ts#L87">services/line.service.ts:87</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -228,7 +228,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/services/LineService.ts#L270">services/LineService.ts:270</a></li>
+									<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/services/LineService.ts#L270">services/line.service.ts:270</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -267,7 +267,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/services/LineService.ts#L63">services/LineService.ts:63</a></li>
+									<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/services/LineService.ts#L63">services/line.service.ts:63</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -303,7 +303,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/services/LineService.ts#L187">services/LineService.ts:187</a></li>
+									<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/services/LineService.ts#L187">services/line.service.ts:187</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -340,7 +340,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/services/LineService.ts#L161">services/LineService.ts:161</a></li>
+									<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/services/LineService.ts#L161">services/line.service.ts:161</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -377,7 +377,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/services/LineService.ts#L40">services/LineService.ts:40</a></li>
+									<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/services/LineService.ts#L40">services/line.service.ts:40</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -414,7 +414,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/services/LineService.ts#L241">services/LineService.ts:241</a></li>
+									<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/services/LineService.ts#L241">services/line.service.ts:241</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -451,7 +451,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/services/LineService.ts#L123">services/LineService.ts:123</a></li>
+									<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/services/LineService.ts#L123">services/line.service.ts:123</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
