+++ docs/modules.html
@@ -118,7 +118,7 @@
 				<section class="tsd-panel tsd-member tsd-kind-property">
 					<a name="graphql" class="tsd-anchor"></a>
 					<h3>graphql</h3>
-					<div class="tsd-signature tsd-kind-icon">graphql<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">GraphQLService</span></div>
+					<div class="tsd-signature tsd-kind-icon">graphql<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">GraphqlService</span></div>
 					<aside class="tsd-sources">
 					</aside>
 				</section>
